fix(ui): extract overlay access lifecycle
